clean-up 2 final touches
Tried to replace the copyright emoji with an image in the footer, but the problem is as the text on its right folds into new lines neither flexbox, nor grid will keep the two together on the right, because the maximum content width of the text is used in both to calculate its space not the space it actually takes after folding.
I dont't know the solution, yet.

>>> PROJECT :   [TYPE PROJECT NAME HERE]